requirements.txt
attrs
discord-py-interactions
motor
pymongo
python-dotenv
